test(mitm-addon): split jsonl writer state tests
