+++ crates/runner/mitm-addon/tests/test_jsonl_writer.py
@@ -1,0 +1,91 @@
+"""Tests for the asynchronous JSONL writer state machine."""
+
+import queue
+import threading
+from unittest.mock import MagicMock, patch
+
+import jsonl_writer
+
+
+def test_full_write_queue_warns_and_does_not_track_dropped_path_state(tmp_path):
+    log_path = str(tmp_path / "net.jsonl")
+    log = MagicMock()
+
+    with (
+        patch.object(jsonl_writer, "_ensure_worker_locked", return_value=True),
+        patch.object(jsonl_writer._queue, "put_nowait", side_effect=queue.Full),
+        patch.object(jsonl_writer.ctx, "log", log, create=True),
+    ):
+        jsonl_writer.write_jsonl_line(log_path, b'{"action":"ALLOW"}\n', "network")
+
+    log.warn.assert_called_once()
+    warning = log.warn.call_args.args[0]
+    assert warning == "Dropping network log because the JSONL writer backlog is full"
+    assert log_path not in jsonl_writer._accepted_by_path
+    assert log_path not in jsonl_writer._completed_by_path
+    assert log_path not in jsonl_writer._flush_waiters_by_path
+    assert not (tmp_path / "net.jsonl").exists()
+
+
+def test_completed_write_prunes_path_state_without_explicit_flush(tmp_path):
+    log_path = str(tmp_path / "proxy.jsonl")
+
+    jsonl_writer.write_jsonl_line(log_path, b'{"message":"done"}\n', "proxy")
+    jsonl_writer._queue.join()
+
+    assert log_path not in jsonl_writer._accepted_by_path
+    assert log_path not in jsonl_writer._completed_by_path
+    assert log_path not in jsonl_writer._flush_waiters_by_path
+    assert (tmp_path / "proxy.jsonl").read_bytes().splitlines()
+
+
+def test_flush_prunes_completed_path_state(tmp_path):
+    log_path = str(tmp_path / "net.jsonl")
+
+    jsonl_writer.write_jsonl_line(log_path, b'{"action":"ALLOW"}\n', "network")
+    jsonl_writer.flush_log_path(log_path)
+
+    assert log_path not in jsonl_writer._accepted_by_path
+    assert log_path not in jsonl_writer._completed_by_path
+
+
+def test_concurrent_flushes_prune_after_all_waiters_complete(tmp_path):
+    log_path = str(tmp_path / "net.jsonl")
+    append_started = threading.Event()
+    release_append = threading.Event()
+    flush_threads: list[threading.Thread] = []
+    original_append_lines = jsonl_writer._append_lines
+
+    def append_lines(path: str, content: bytes) -> None:
+        append_started.set()
+        release_append.wait()
+        original_append_lines(path, content)
+
+    with patch.object(jsonl_writer, "_append_lines", side_effect=append_lines):
+        try:
+            jsonl_writer.write_jsonl_line(log_path, b'{"action":"ALLOW"}\n', "network")
+            assert append_started.wait(timeout=1)
+
+            flush_threads = [
+                threading.Thread(target=jsonl_writer.flush_log_path, args=(log_path,)),
+                threading.Thread(target=jsonl_writer.flush_log_path, args=(log_path,)),
+            ]
+            for thread in flush_threads:
+                thread.start()
+
+            with jsonl_writer._condition:
+                assert jsonl_writer._condition.wait_for(
+                    lambda: jsonl_writer._flush_waiters_by_path.get(log_path, 0) == 2,
+                    timeout=1,
+                )
+        finally:
+            release_append.set()
+            for thread in flush_threads:
+                thread.join(timeout=1)
+
+    for thread in flush_threads:
+        assert not thread.is_alive()
+
+    assert log_path not in jsonl_writer._accepted_by_path
+    assert log_path not in jsonl_writer._completed_by_path
+    assert log_path not in jsonl_writer._flush_waiters_by_path
